Update luks-on-lvm bootloader command again
I feel like I missed, or incompletely logged, several previous changes.
The /boot partition is on /dev/nvme*, not /dev/sda*, and has been there
for a while I suspect; I think the old UUID might have been outdated
too. Be that as it may, the UUID definitely changed again just now, and
I’ve also taken the opportunity to drop the deprecated directive for the
unified cgroup hierarchy – that’s now the default both in upstream
systemd and in the Arch package.